+++ __init__.py
@@ -1,4 +1,9 @@
-from .loaders import CoreMLLoaderUNet
+import os
+import sys
+
+sys.path.append(os.path.dirname(__file__))
+
+from coreml_suite import CoreMLLoaderUNet
 
 NODE_CLASS_MAPPINGS = {
     "CoreMLUNetLoader": CoreMLLoaderUNet,
