+++ coreml_suite/__init__.py
@@ -1,0 +1,3 @@
+from .loaders import CoreMLLoaderUNet
+
+__all__ = ["CoreMLLoaderUNet"]
